move validations tests to test_validation
